#11 refactor: projectMember 엔티티 수정
status를 enum 클래스로 관리 및 거절 혹은 강퇴 시 사유를 적을 수 있도록 comment 칼럼 추가
