+++ dev-together/src/main/java/com/project/devtogether/participant/domain/ProjectMember.java
@@ -1,7 +1,12 @@
-package com.project.devtogether.project.domain;
+package com.project.devtogether.participant.domain;
 
 import com.project.devtogether.member.domain.Member;
+import com.project.devtogether.participant.domain.enums.ParticipantStatus;
+import com.project.devtogether.project.domain.Project;
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -30,11 +35,19 @@ public class ProjectMember {
     @JoinColumn(name = "member_id")
     private Member member;
 
+    @Column(length = 25, nullable = false)
+    @Enumerated(EnumType.STRING)
+    private ParticipantStatus participantStatus;
+
+    @Column(length = 500)
+    private String comment;
+
     protected ProjectMember() {}
 
     private ProjectMember(Project project, Member member) {
         setProject(project);
         this.member = member;
+        this.participantStatus = ParticipantStatus.APPLIED;
     }
 
     public static ProjectMember of(Project project, Member member) {
@@ -49,3 +62,4 @@ public class ProjectMember {
         project.getMembers().add(this);
     }
 }
+
