+++ dev-together/src/main/java/com/project/devtogether/participant/domain/enums/ParticipantStatus.java
@@ -1,0 +1,8 @@
+package com.project.devtogether.participant.domain.enums;
+
+public enum ParticipantStatus {
+    APPLIED, //신청
+    ACCEPTED, //승인
+    REJECTED, //거부
+    EXPELLED //강퇴
+}
